glProjection: Doesn't use glGet*(GL_MODELVIEW_MATRIX), i.e works with OpenGL ES 2.0 GLvector uses GLfloat

+++ lib/gpumisc/GLvector.h
@@ -4,6 +4,6 @@
 #include "gl.h"
 
 #include "tvector.h"
-typedef tvector<3,GLdouble> GLvector;
+typedef tvector<3,GLfloat> GLvector;
 
 #endif // GLVECTOR_H
